LayoutTests/fast/dom/Window/area-rel-noopener-expected.txt

CONSOLE MESSAGE: line 11: PASS: window.opener is null
Test that window.opener is null when a new window is opened from an area element with rel='noopener'.
